--- src/App.css
@@ -1,126 +1,0 @@
-
-#root {
-  max-width: 1920px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
-
-/* //stile für cart   */
-.cart-page {
-  display: flex;
-  gap: 2rem;
-  margin: 2rem auto;
-  max-width: 1100px;
-}
-.cart-list {
-  flex: 2;
-}
-.cart-summary {
-  flex: 1;
-  background: #f8fae7;
-  padding: 2rem;
-  border-radius: 12px;
-  min-width: 320px;
-}
-.cart-item {
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-  background: #fff;
-  margin-bottom: 1rem;
-  border-radius: 10px;
-  padding: 1rem;
-  box-shadow: 0 2px 6px #eee;
-}
-.cart-item__img {
-  width: 80px;
-  height: 80px;
-  object-fit: contain;
-  border-radius: 8px;
-}
-.cart-item__info {
-  flex: 1;
-}
-.cart-item__name {
-  font-weight: 500;
-  margin-bottom: 0.6rem;
-}
-.cart-item__qty-controls button {
-  width: 32px;
-  height: 32px;
-  font-size: 1.3rem;
-}
-.cart-item__price .old {
-  margin-left: 0.5rem;
-  color: #999;
-  text-decoration: line-through;
-}
-.cart-item__remove {
-  background: none;
-  border: none;
-  font-size: 2rem;
-  color: #888;
-  cursor: pointer;
-}
-.cart-form input {
-  display: block;
-  margin-bottom: 0.7rem;
-  width: 100%;
-  padding: 0.7rem;
-  border-radius: 6px;
-  border: 1px solid #ddd;
-}
-.cart-form button {
-  width: 100%;
-  background: #a4b94c;
-  color: #fff;
-  padding: 0.8rem;
-  border-radius: 6px;
-  border: none;
-  font-weight: 600;
-  font-size: 1rem;
-  margin-top: 0.7rem;
-}
-.form-error {
-  color: #d33434;
-  font-size: 0.92rem;
-  margin-bottom: 0.6rem;
-  display: block;
-}
