+++ src/components/Categories/Categories.scss
@@ -1,124 +1,129 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: "Montserrat", serif;
-  line-height: 1.5;
+// .categories-container {
+//   max-width: 1500px;
+//   margin: 0 auto;
 
-  
-}
-body{
-   background: #FFFFF1; 
-}
+// }
 
-.categories-container {
-  max-width: 1500px;
-  margin: 0 auto;
+// .categories-section {
+//   width: 100%;
+//   background: #FFFFF1;
+//   font-family: 'Montserrat', sans-serif;
+//   max-width: 1360px;
+//   padding-bottom: 80px;
 
-}
+//   .categories-title {
+//     height: 70px;
+//     margin: 40px 0;
+//     display: flex;
+//     align-items: center;
+//     justify-content: center;
 
-.categories-section {
-  width: 100%;
-  background: #FFFFF1;
-  font-family: 'Montserrat', sans-serif;
-  max-width: 1360px;
-  padding-bottom: 80px;
-  
-  
+//     h2 {
+//       font-size: 48px;
+//       font-weight: bold;
+//       color: #424436;
+//       margin: 0;
+//       text-align: center;
+//     }
+//   }
 
-  .categories-title {
-    height: 70px;
-    margin: 40px 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+//   .categories-grid {
+//     display: grid;
+//     grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
+//     gap: 32px;
+//     width: 100%;
+//     padding-bottom: 40px;
 
-    h2 {
-      font-size: 48px;
-      font-weight: bold;
-      color: #424436;
-      margin: 0;
-      text-align: center;
-    }
-  }
+//     .category-card {
+//       background: white;
+//       border-radius: 24px;
+//       box-shadow: 0 2px 12px #2221;
+//       display: flex;
+//       flex-direction: column;
+//       align-items: center;
+//       transition: box-shadow 0.2s ease, transform 0.2s ease;
+//       cursor: pointer;
+//       background: #FFFFF1;
 
-  .categories-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-    gap: 32px;
+//       &:hover {
+//         box-shadow: 0 4px 18px #2222;
+//         transform: translateY(-4px);
+//       }
+
+//       .category-image {
+//         width: 100%;
+//         height: 220px;
+//         border-radius: 24px 24px 0 0;
+//         object-fit: cover;
+//         object-position: center;
+//         max-width: 100%;
+//       }
+
+//       .category-label {
+//         width: 100%;
+//         padding: 14px 0;
+//         color: #424436;
+//         text-align: center;
+//         font-size: 20px;
+//         font-weight: 500;
+//         line-height: 26px;
+//         transition: color 0.2s ease;
+//       }
+//     }
+//   }
+// }
+
+// @media (max-width: 900px) {
+//   .categories-container {
+//     padding: 0 18px;
+//   }
+
+//   .categories-title h2 {
+//     font-size: 32px;
+//   }
+
+//   .categories-grid {
+//     gap: 18px;
+//   }
+// }
+
+// @media (max-width: 600px) {
+//   .categories-container {
+//     padding: 0 12px;
+//   }
+
+//   .categories-title h2 {
+//     font-size: 24px;
+//   }
+
+//   .categories-grid {
+//     gap: 10px;
+//     padding-bottom: 18px;
+//   }
+// }
+
+// @media (max-width: 400px) {
+//   .categories-grid {
+//     grid-template-columns: 1fr;
+//   }
+// }
+
+
+
+
+.category-card {
+  text-decoration: none;
+
+  &__image {
     width: 100%;
-    padding-bottom: 40px;
-
-    .category-card {
-      background: white;
-      border-radius: 24px;
-      box-shadow: 0 2px 12px #2221;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      transition: box-shadow 0.2s ease, transform 0.2s ease;
-      cursor: pointer;
-      background: #FFFFF1; 
-
-      &:hover {
-        box-shadow: 0 4px 18px #2222;
-        transform: translateY(-4px);
-      }
-
-      .category-image {
-        width: 100%;
-        height: 220px;
-        border-radius: 24px 24px 0 0;
-        object-fit: cover;
-        object-position: center;
-        max-width: 100%;
-      }
-
-      .category-label {
-        width: 100%;
-        padding: 14px 0;
-        color: #424436;
-        text-align: center;
-        font-size: 20px;
-        font-weight: 500;
-        line-height: 26px;
-        transition: color 0.2s ease;
-      }
-    }
-  }
-}
-
-@media (max-width: 900px) {
-  .categories-container {
-    padding: 0 18px;
+    max-height: 200px;
+    height: 100%;
+    display: block;
+    object-fit: cover;
+    border-radius: 14px;
   }
 
-  .categories-title h2 {
-    font-size: 32px;
-  }
-
-  .categories-grid {
-    gap: 18px;
-  }
-}
-
-@media (max-width: 600px) {
-  .categories-container {
-    padding: 0 12px;
-  }
-
-  .categories-title h2 {
-    font-size: 24px;
-  }
-
-  .categories-grid {
-    gap: 10px;
-    padding-bottom: 18px;
-  }
-}
-
-@media (max-width: 400px) {
-  .categories-grid {
-    grid-template-columns: 1fr;
+  &__label {
+    color: #111;
   }
 }
