+++ src/components/Hero/Hero.scss
@@ -1,107 +1,67 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: "Montserrat", serif;
-  line-height: 1.5;
-}
 
-.hero-section {
-  position: relative;
-  width: 100%;
-  min-height: 400px;
-  height: 60vh;              
-  max-height: 700px;
-  overflow: hidden;
-  background: #FFFFF1;
-  display: flex;
-  align-items: stretch;
-  justify-content: center;
-}
-
-.hero-image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  display: block;
-
-}
-
-.hero-content {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  padding: 3vw 5vw;          
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: flex-start;
-  box-sizing: border-box;
-  background: linear-gradient(90deg, rgba(0,0,0,0.38) 0%, rgba(0,0,0,0.01) 90%); 
-}
-
-.hero-text {
-  color: white;
-  font-family: 'Montserrat', sans-serif;
-  font-weight: 700;
-  font-size: clamp(2rem, 6vw, 5.5rem); 
-  line-height: 1.1;
-  margin-bottom: 24px;
-  max-width: 900px;
-  text-shadow: 0 4px 24px #2228;
-}
-
-.hero-button {
-  width: auto;
-  min-width: 150px;
-  height: auto;
-  padding: 16px 40px;
-  border-radius: 8px;
-  background-color: #92A134;
-  color: white;
-  font-family: 'Montserrat', sans-serif;
-  font-size: 1.1rem;
-  font-weight: 600;
-  border: none;
-  cursor: pointer;
-  margin-top: 18px;
-  box-shadow: 0 2px 12px #2222;
-  transition: background 0.2s;
-}
-.hero-button:hover {
-  background-color: #72851c;
-}
-
-
-@media (max-width: 900px) {
-  .hero-text {
-    font-size: clamp(1.4rem, 6vw, 3rem);
-    max-width: 90vw;
+.hero {
+  &-section {
+    position: relative;
+    width: 100%;
+    max-height: 600px;
+    overflow: hidden;
+    background: #fffff1;
+    max-width: 1440px;
+    margin: 0 auto;
   }
-  .hero-content {
-    padding: 8vw 5vw;
-  }
-}
 
-@media (max-width: 600px) {
-  .hero-section {
-    height: 40vh;
-    min-height: 180px;
+  &-image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    display: block;
   }
-  .hero-content {
-    padding: 4vw 3vw;
+
+  &-content {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
     align-items: flex-start;
+    box-sizing: border-box;
+    background: linear-gradient(
+      90deg,
+      rgba(0, 0, 0, 0.38) 0%,
+      rgba(0, 0, 0, 0.01) 90%
+    );
+    padding: 24px;
   }
-  .hero-text {
-    font-size: 1.3rem;
-    margin-bottom: 10px;
-    max-width: 98vw;
+
+  &-text {
+    color: white;
+    font-family: "Montserrat", sans-serif;
+    font-weight: 700;
+    font-size: clamp(2rem, 6vw, 5.5rem);
+    line-height: 1.1;
+    margin-bottom: 24px;
+    max-width: 900px;
+    text-shadow: 0 4px 24px #2228;
   }
-  .hero-button {
-    padding: 12px 24px;
-    font-size: 1rem;
-    margin-top: 12px;
+
+  &-button {
+    border-radius: 8px;
+    background-color: #92a134;
+    color: white;
+    font-family: "Montserrat", sans-serif;
+    font-size: 16px;
+    font-weight: 600;
+    border: none;
+    cursor: pointer;
+    box-shadow: 0 2px 12px rgba(34, 34, 34, 0.1333333333);
+    transition: background 0.2s;
+    padding: 16px 54px;
+
+    &:hover {
+      background-color: #72851c;
+    }
   }
 }
