+++ src/components/SalesItems/SaleItems.scss
@@ -1,9 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&display=swap');
 
-body {
-  font-family: 'Montserrat', sans-serif;
-}
-
 .saleitems-outer-container {
   max-width: 1500px;
   margin: 0 auto;
