+++ src/components/ToolsAndEquipment/ToolsAndEquipment.scss
@@ -1,8 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&display=swap');
 
-body {
-  font-family: 'Montserrat', sans-serif;
-}
 
 .items-container {
   width: 1360px;
