+++ src/main.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
+import "./index.scss"
 
 ReactDOM.createRoot(document.getElementById("root")).render(<App />);
