--- src/pages/Categories.jsx
@@ -1,14 +1,0 @@
-// src/pages/Categories.jsx
-import React from "react";
-
-const Categories = () => {
-  return (
-    <div style={{ padding: "2rem" }}>
-      <h1>Categories</h1>
-      <p>Hier siehst du alle Produktkategorien.</p>
-      {/* Hier kommt deine bestehende Kategorien-Logik hin */}
-    </div>
-  );
-};
-
-export default Categories;
