--- src/pages/HomePage.jsx
@@ -1,16 +1,0 @@
-import Header from "../components/Header";
-import Banner from "../components/Banner";
-import Categories from "../components/Categories";
-import Footer from "../components/Footer";
-
-function HomePage() {
-  return (
-    <>
-      <Header />
-      <Banner />
-      <Categories />
-      <Footer />
-    </>
-  );
-}
-export default HomePage;
